introduce uhubctl.h for defines and struct definition
To improve readability of the code move the defines and struct
definitons to a separate header file. Furthermore add this headerfile to
the Makefile.
